+++ src/document/url_line_element.cpp
@@ -5,6 +5,11 @@ UrlLineElement::UrlLineElement(const std::string & content) :
 	_url(content.substr(content.find_first_of("(")+1, content.length()-1))
 {}
 
+UrlLineElement::UrlLineElement(const std::string & content, const std::string & url) :
+	LineElement(content),
+	_url(url)
+{}
+
 std::string UrlLineElement::url(void) const {
 	return _url;
 }